pdf_table_augmenter_api/requirements.txt
Django~=5.2.1
djangorestframework==3.16.0
psycopg2==2.9.10
django-environ==0.12.0
pdfplumber==0.11.7
openai~=1.68.2
django-cors-headers==4.7.0
yake~=0.6.0
langdetect==1.0.9
docling==2.39.0
roman==5.1